Restructure directories: move WorldPos to Component/, rename Scene/ to Phase/, close #31
