+++ Ash2/Ash2.vcxproj.filters
@@ -9,10 +9,10 @@
       <UniqueIdentifier>{93995380-89BD-4b04-88EB-625FBE52EBFB}</UniqueIdentifier>
       <Extensions>h;hh;hpp;hxx;hm;inl;inc;xsd</Extensions>
     </Filter>
-    <Filter Include="Header Files\Scene">
+    <Filter Include="Header Files\Phase">
       <UniqueIdentifier>{c1d2e3f4-a5b6-7c8d-9e0f-1a2b3c4d5e6f}</UniqueIdentifier>
     </Filter>
-    <Filter Include="Source Files\Scene">
+    <Filter Include="Source Files\Phase">
       <UniqueIdentifier>{d2e3f4a5-b6c7-8d9e-0f1a-2b3c4d5e6f7a}</UniqueIdentifier>
     </Filter>
     <Filter Include="Header Files\Component">
@@ -174,8 +174,8 @@
     <ClCompile Include="Config\ScenarioData.cpp">
       <Filter>Source Files\Config</Filter>
     </ClCompile>
-    <ClCompile Include="Scene\ScenarioPhase.cpp">
-      <Filter>Source Files\Scene</Filter>
+    <ClCompile Include="Phase\ScenarioPhase.cpp">
+      <Filter>Source Files\Phase</Filter>
     </ClCompile>
     <ClCompile Include="System\DrawSystem.cpp">
       <Filter>Source Files\System</Filter>
@@ -183,11 +183,11 @@
     <ClCompile Include="Main.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
-    <ClCompile Include="Scene\DemoPhase.cpp">
-      <Filter>Source Files\Scene</Filter>
+    <ClCompile Include="Phase\DemoPhase.cpp">
+      <Filter>Source Files\Phase</Filter>
     </ClCompile>
-    <ClCompile Include="Scene\PhaseStack.cpp">
-      <Filter>Source Files\Scene</Filter>
+    <ClCompile Include="Phase\PhaseStack.cpp">
+      <Filter>Source Files\Phase</Filter>
     </ClCompile>
     <ClCompile Include="stdafx.cpp">
       <Filter>Source Files</Filter>
@@ -884,23 +884,23 @@
     <ClInclude Include="Config\ScenarioData.hpp">
       <Filter>Header Files\Config</Filter>
     </ClInclude>
-    <ClInclude Include="Scene\ScenarioPhase.hpp">
-      <Filter>Header Files\Scene</Filter>
+    <ClInclude Include="Phase\ScenarioPhase.hpp">
+      <Filter>Header Files\Phase</Filter>
     </ClInclude>
     <ClInclude Include="stdafx.h">
       <Filter>Header Files</Filter>
     </ClInclude>
-    <ClInclude Include="WorldPos.hpp">
-      <Filter>Header Files</Filter>
+    <ClInclude Include="Component\WorldPos.hpp">
+      <Filter>Header Files\Component</Filter>
     </ClInclude>
-    <ClInclude Include="Scene\DemoPhase.hpp">
-      <Filter>Header Files\Scene</Filter>
+    <ClInclude Include="Phase\DemoPhase.hpp">
+      <Filter>Header Files\Phase</Filter>
     </ClInclude>
-    <ClInclude Include="Scene\IPhase.hpp">
-      <Filter>Header Files\Scene</Filter>
+    <ClInclude Include="Phase\IPhase.hpp">
+      <Filter>Header Files\Phase</Filter>
     </ClInclude>
-    <ClInclude Include="Scene\PhaseStack.hpp">
-      <Filter>Header Files\Scene</Filter>
+    <ClInclude Include="Phase\PhaseStack.hpp">
+      <Filter>Header Files\Phase</Filter>
     </ClInclude>
     <ClInclude Include="ThirdParty\entt\entt.hpp">
       <Filter>ThirdParty\entt</Filter>
